Add a simple test
Adds rack-test and minitest for testing.

By using prepend in test_helper.rb, we're making sure that we'll get the
local one in the current directory instead of the one we installed as a
gem when we use require "ray".

We're also using expand_path to convert the relative path to absolute.
That's important if anything might change the current directory.