Add autoloading
Now, when the app attempts to use a constant that is not explicitly
required, it'll first require the corresponding file and then return the
constant.

Autoloading in Rails is way more comprehensive than this, but we don't
need all that functionality now. Check out the source if needed:

activesupport/lib/active_support/dependencies.rb

To convert the constant to a file name in snake case, a utility is used.
The method definition is copied from ActiveSupport underscore method.
But it is not the latest version of the method.