Render default action view when render is not present
Previously a controller action would look like this:

def show
  @model = Model.find(params["id"])

  render :show
end

The current version of Ray depended on the return value of the app
controller action (show method in the example). So if render :show line
is removed, the return value changes and the action would not work.

The new one acts like Rails. If the action does not have a render method
call, Ray automatically calls the default view of that action:

def show
  @model = Model.find(params["id"])

  # render method is not called, will render show.html.erb
end

Since the return value of controller action is now ignored, you can do
other things after the render method call:

def show
  @model = Model.find(params["id"])

  render :show

  puts 'Hello!'
end

It would not cause an error and the view would render as usual.