Separate view from controller
Views can have methods for escaping strings or JavaScript, generating
HTML tags like buttons, links and forms. And these kind of functionality
should not be present in the controller object.

Check out erubis user guide to understand how it works. erubis
recommends using evaluate instead of result. The relevant keyword in the
documentation is "context":

http://www.kuwata-lab.com/erubis/users-guide.html

I also added an example helper method for escaping strings to see how it
works. It is similar to:

https://api.rubyonrails.org/classes/ERB/Util.html#method-c-h

Example usage is as follows:

Template:

<p>
  View helper test: <%= h "http://example.com/?a=\11\15" %>
</p>

Output:

<p>
  View helper test: http://example.com/?a=%09%0D
</p>