Add sqlite support
For now, Ray can not learn the database name from the application on
run-time, that functionality will be added later on. In the meantime,
we'll assume the name of the database is "ray".

This commit adds methods that allow us to inspect the schema.

Assuming the application has the following migration in the migration.rb
file:

require "sqlite3"

db = SQLite3::Database.new "ray.db"

db.execute <<-SQL
  CREATE TABLE my_table (
    id INTEGER PRIMARY KEY,
    posted INTEGER,
    title VARCHAR(30),
    body VARCHAR(32000)
  );
SQL

When "bundle exec ruby migration.rb" is run, a database called ray and a
table called my_table will be created.

When the following model is defined:

class MyTable < Ray::Model::SQLite
end

The app can inspect the database schema as follows:

puts MyTable.schema.inspect
=> {"id"=>"INTEGER", "posted"=>"INTEGER", "title"=>"VARCHAR(30)", "body"=>"VARCHAR(32000)"}

Relevant links:

https://github.com/sparklemotion/sqlite3-ruby/blob/master/lib/sqlite3/pragmas.rb#L527
https://sqlite.org/pragma.html#pragma_table_info