Add	create method
We convert the values to their respective SQLite data types before
inserting.

last_insert_rowid() is used to get the id of the last inserted record.

https://www.sqlite.org/c3ref/last_insert_rowid.html

After insertion, we return an object that represents the row.

The method could be used in an app as follows:

record = MyTable.create("title" => "It happened!", "posted" => 1, "body" => "It did!")
puts record.inspect