Add column accessors
Currently we're accessing the columns as follows:

record = MyTable.find(id)
puts record["title"]
record["title"] = "Hello world!"

But Rails adds accessor methods so that we can do the following:

record = MyTable.find(id)
puts record.title
record.title = "Hello world!"

This commit adds an accessor method only if we need it. Initially no
accessor methods are defined. When an accessor method is used,
method_missing is called. In method_missing, the accessor method is
added and the value of that column is returned (for reader). If that
accessor is used again, method_missing is not called since we already
defined a method for it. The writer works similarly.