Implement a rudimentary routing DSL
This commit adds a match method that allows applications to declare the
routes.

This match method is similar to the one in Rails. It is the most
versatile but the most complicated way to specify a route. But other
routing methods like root, get, post and resources can be implemented on
top of it.

In this implementation, every action on every controller acts like a
Rack app.

- The action class method and dispatch instance method on controller
  represents usual route -> controller -> action flow.
- The first line of the get_destination method (if destination responds
  to a call method) represents the routing of a request to another rack
  app.

On an incoming request, the application gets the rack app and calls the
call method on it.

Here is an example route configuration and the corresponding paths:

app = BestQuotes::Application.new

app.route do
  # matches /sub-app, renders "Hello, sub-app!"
  match "sub-app", proc { [200, {}, ["Hello, sub-app!"]]}

  # matches /quotes/2, renders quotes#show action
  match ":controller/:id", :default => { "action" => "show" }

  # matches /quotes, renders quotes#index action
  match ":controller", :default => { "action" => "index" }

  # matches /quotes/a_quote, renders quotes#a_quote action
  match "quotes/a_quote", "quotes#a_quote"
end

run app