Add a method to get the running app instance
When an app is booted up, we'll store the current app instance in a
variable. This will allow us to use the following syntax to it:

Ray.application

This is again similar to Ruby on Rails. It will give us a better way to
configure an app on boot time. An example for route configuration using
this method:

Ray.application.route do
  ...
end